pyproject.toml
[project]
name = "learn-ai"
version = "0.1.0"
description = "Personal AI knowledge base"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12",
    "feedparser>=6.0.12",
    "httpx>=0.27",
    "mkdocs-material>=9.7.6",
    "pyyaml>=6.0",
]
